--- Bases web/Convenios/index.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Relatório de Gastos</title>
-    <link rel="stylesheet" href="styles.css">
-    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-</head>
-<body>
-    <div class="container">
-        <input type="text" id="search" placeholder="Pesquise por interessado...">
-        <div class="controls">
-            <button id="toggle-view">Visualização Diária</button>
-        </div>
-        <canvas id="chart"></canvas>
-        <table id="data-table">
-            <thead>
-                <tr>
-                    <th>Data</th>
-                    <th>Total Gasto (R$)</th>
-                    <th>Interessado</th>
-                </tr>
-            </thead>
-            <tbody></tbody>
-        </table>
-    </div>
-    <script src="script.js"></script>
-</body>
-</html>
