--- Bases web/Convenios/styles.css
@@ -1,72 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    background-color: #f4f4f4;
-    margin: 0;
-    padding: 0;
-    color: #333;
-}
-
-.container {
-    width: 80%;
-    margin: auto;
-    padding: 20px;
-    background-color: #fff;
-    border-radius: 8px;
-    box-shadow: 0 0 10px rgba(0,0,0,0.1);
-}
-
-#search {
-    width: 100%;
-    padding: 10px;
-    margin-bottom: 20px;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    font-size: 16px;
-}
-
-.controls {
-    margin-bottom: 20px;
-}
-
-#toggle-view {
-    padding: 10px 20px;
-    border: none;
-    border-radius: 4px;
-    background-color: #007bff;
-    color: #fff;
-    font-size: 16px;
-    cursor: pointer;
-}
-
-#toggle-view:hover {
-    background-color: #0056b3;
-}
-
-canvas {
-    max-width: 100%;
-    height: auto;
-    margin-bottom: 20px;
-}
-
-table {
-    width: 100%;
-    border-collapse: collapse;
-}
-
-table th, table td {
-    border: 1px solid #ddd;
-    padding: 8px;
-    text-align: right;
-}
-
-table th {
-    background-color: #f2f2f2;
-}
-
-table td {
-    text-align: center;
-}
-
-table .data-date {
-    text-align: left;
-}
