--- Bases web/Verbas de escola/style.css
@@ -1,50 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    background-color: #f4f4f4;
-    margin: 0;
-    padding: 20px;
-}
-
-.container {
-    max-width: 600px;
-    margin: 0 auto;
-    background: white;
-    padding: 20px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-}
-
-h1 {
-    text-align: center;
-    color: #333;
-}
-
-input[type="text"] {
-    width: 100%;
-    padding: 10px;
-    margin-bottom: 10px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-}
-
-#suggestions {
-    border: 1px solid #ccc;
-    border-top: none;
-    max-height: 150px;
-    overflow-y: auto;
-    background: white;
-}
-
-#suggestions div {
-    padding: 10px;
-    cursor: pointer;
-}
-
-#suggestions div:hover {
-    background-color: #f0f0f0;
-}
-
-canvas {
-    width: 100%;
-    max-width: 600px;
-    height: 300px;
-}
