+++ web/index.html
@@ -13,29 +13,39 @@
 <body>
 	<div class="main">
 		<!-- Navbar -->
-		<div class="navbar">
-			<div class="menu">
-				<ul>
-					<img id="logo_gastos" src="imagens/favicon.png" />
-					<li>
-						<a href="index.html">
-							<h2 id="botao1">Início</h2>
-						</a>
-					</li>
-					<li>
-						<a href="sobre_nos.html">
-							<h2 id="botao1">Sobre nós</h2>
-						</a>
-					</li>
-					<li>
-						<a href="sobre_projeto.html">
-							<h2 id="botao1">Sobre o projeto</h2>
-						</a>
-					</li>
-				</ul>
-			</div>
+	<div class="navbar">
+		<div class="menu">
+			<ul>
+				<img id="logo_gastos" src="imagens/favicon.png" />
+				<li>
+					<a href="index.html">
+						<h2 id="botao1">Início</h2>
+					</a>
+				</li>
+				<li>
+					<a href="sobre_nos.html">
+						<h2 id="botao1">Sobre nós</h2>
+					</a>
+				</li>
+				<li>
+					<a href="sobre_projeto.html">
+						<h2 id="botao1">Sobre o projeto</h2>
+					</a>
+				</li>
+				<li class="dropdown">
+					<a href="javascript:void(0)">
+						<h2 id="botao1">Gráficos</h2>
+					</a>
+					<div class="dropdown-content">
+						<a href="Bases web/Convenios/index_convenios.html">Convênios</a>
+						<a href="opcao2.html">Opção 2</a>
+						<a href="opcao3.html">Opção 3</a>
+					</div>
+				</li>
+			</ul>
 		</div>
-
+	</div>
+</div>
 		<!-- Logo -->
 		<div class="logo_container">
 			<img src="imagens/simbolo.png" alt="Logo do gastos DF" />
